Rename gemini-generator-> text, nano-generator -> nano
