+++ packages/gemini-kit/src/components/text.ts
